+++ lib/web_ui/test/canvaskit/canvaskit_api_test.dart
@@ -1311,7 +1311,7 @@ void _canvasTests() {
         devicePixelRatio * kLightRadius,
         tonalColors.ambient,
         tonalColors.spot,
-        flags,
+        flags.toDouble(),
       );
     }
   });
@@ -1697,16 +1697,16 @@ void _paragraphTests() {
 
     // "Hello"
     for (int i = 0; i < 5; i++) {
-      expect(paragraph.getWordBoundary(i).start, 0);
-      expect(paragraph.getWordBoundary(i).end, 5);
+      expect(paragraph.getWordBoundary(i.toDouble()).start, 0);
+      expect(paragraph.getWordBoundary(i.toDouble()).end, 5);
     }
     // Placeholder
     expect(paragraph.getWordBoundary(5).start, 5);
     expect(paragraph.getWordBoundary(5).end, 6);
     // "World"
     for (int i = 6; i < 11; i++) {
-      expect(paragraph.getWordBoundary(i).start, 6);
-      expect(paragraph.getWordBoundary(i).end, 11);
+      expect(paragraph.getWordBoundary(i.toDouble()).start, 6);
+      expect(paragraph.getWordBoundary(i.toDouble()).end, 11);
     }
     // "!"
     expect(paragraph.getWordBoundary(11).start, 11);
@@ -1788,13 +1788,13 @@ void _paragraphTests() {
     final int sampleCount = gl.getParameter(gl.samples);
     final int stencilBits = gl.getParameter(gl.stencilBits);
 
-    final int glContext = canvasKit.GetWebGLContext(
+    final double glContext = canvasKit.GetWebGLContext(
       canvas,
       SkWebGLContextOptions(
         antialias: 0,
         majorVersion: webGLVersion.toDouble(),
       ),
-    ).toInt();
+    );
     final SkGrContext grContext =  canvasKit.MakeGrContext(glContext);
     final SkSurface? skSurface = canvasKit.MakeOnScreenGLSurface(
       grContext,
